add GetItem test and change CreateItem test

+++ lesson-1/shop/models/item.go
@@ -1,7 +1,7 @@
 package models
 
 type Item struct {
-	ID    int32   `json:"id"`
-	Name  string  `json:"name"`
-	Price float64 `json:"price"`
+	ID    int32  `json:"id"`
+	Name  string `json:"name"`
+	Price int32  `json:"price"`
 }
